0. Innovation and Product Development/codes/Website_Christoffer/requirements.txt
 flask==0.9
 flask-mail==0.7.6
 sqlalchemy==0.7.9
 flask-sqlalchemy==0.16
 sqlalchemy-migrate==0.7.2
 flask-whooshalchemy==0.55a
 flask-wtf==0.8.4
 pytz==2013b
 flask-babel==0.8
 flup
 pydocumentdb>=1.0.0
 azure==0.8.4
 mysql_connector_python==8.0.6
 numpy==1.14.2
 turbogears2==2.3.12
 kajiki==0.7.2
 tgext.webassets==0.0.2
 dukpy==0.2.0
 